requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
Werkzeug>=2.3.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0